requirements_live.txt
pandas>=1.5.0
numpy>=1.21.0
requests>=2.28.0
websocket-client>=1.4.0
